update contact form and extract UI
